requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.12.0
torchvision==0.13.0

# numpy==1.23.5
Pillow==9.3.0
matplotlib==3.6.2
absl-py==2.1.0
imageio==2.27.0
pandas==1.5.3
scikit-image==0.22.0
scikit-learn==1.2.2
opencv-python==4.9.0.80
scipy==1.10.1
seaborn==0.13.0
PyYAML==6.0
albumentations==1.4.0
h5py==3.10.0
wandb==0.16.6

